requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Data handling
pandas>=1.3.0
pyyaml>=5.4.0
tqdm>=4.60.0
statsmodels>=0.13.0

# Analysis tools
umap-learn>=0.5.0

# Configuration management
hydra-core>=1.1.0
omegaconf>=2.1.0

# Optional dependencies
# Install with: pip install -r requirements-dev.txt
